[2.4.3] 更新和修复
