+++ gradle.properties
@@ -1,5 +1,5 @@
 group=me.arasple.mc.trchat
-version=2.4.2
+version=2.4.3
 kotlin.incremental=true
 kotlin.incremental.java=true
 kotlin.caching.enabled=true
