+++ project/runtime-bukkit/src/main/kotlin/me/arasple/mc/trchat/module/display/function/standard/Mention.kt
@@ -6,10 +6,7 @@ import me.arasple.mc.trchat.module.conf.file.Functions
 import me.arasple.mc.trchat.module.display.function.Function
 import me.arasple.mc.trchat.module.display.function.StandardFunction
 import me.arasple.mc.trchat.module.internal.script.Reaction
-import me.arasple.mc.trchat.util.CooldownType
-import me.arasple.mc.trchat.util.isInCooldown
-import me.arasple.mc.trchat.util.passPermission
-import me.arasple.mc.trchat.util.updateCooldown
+import me.arasple.mc.trchat.util.*
 import org.bukkit.entity.Player
 import taboolib.common.platform.Platform
 import taboolib.common.platform.PlatformSide
