+++ project/runtime-bukkit/src/main/resources/lang/en_US.yml
@@ -155,6 +155,10 @@ Function-Mention-All-Format:
     args: [ ]
 
 Private-Message-Spy-Format: '&8[&3Spy&8] &6{0} &2-> &3{1}&f: &7{2}'
+Private-Message-Spy-Format-New:
+  - type: json
+    text: '&8\[&3Spy&8\] &6{0} &2-> &3{1}&f: [message]'
+    args: [ ]
 Private-Message-Spy-On:
   - type: actionbar
     text: '&c&l&k|&f &a&lSpy Mode Enabled &c&l&k|'
